Fix run-on thread after app close
The AsyncTask in MainActivity handles the persistent socket connection and reconnection logic. Currently, there are issues with the app where a traditional press of the back button on the Android phone terminates the activity but not the AsyncTask. Starting the app again results in failure to re-establish socket connection and communicate with the alarm system.

Solve these issues by performing status checks of the AsyncTask in the onPause, onStop, and onDestroy lifecycle processes and cancel the task if it is alive at those points.